webui: restructure Attachments/Preview/Current/Item/ to match flat naming scheme
Following the ChatAttachments flattening in
5231f1ae71, the 'Preview/CurrentItem/' directory
containing 'CurrentItem.svelte' was redundant --
the camelCase name reads as 'Current/Item' when
split on capital letters.

Restructured to:
  Attachments/Preview/Current/Item/
    Item.svelte    (was CurrentItem.svelte)
    Audio.svelte
    Image.svelte
    Pdf.svelte
    Text.svelte
    Unavailable.svelte
    Video.svelte

Updated the barrel export in chat/index.ts to
point to the new path.
