+++ examples/server/webui_llamacpp/src/lib/components/app/chat/Attachments/Preview/Current/Item/Text.svelte
